feat(api): change type of some props in requests/responses(DTO)

+++ API/DTO/ProductNutrition/ProductNutritionResponse.cs
@@ -5,9 +5,8 @@ public class ProductNutritionResponse
     public int Id { get; set; }
     public string Name { get; set; }
     public int Calories { get; set; }
-    public int Protein { get; set; }
-    public int Fat { get; set; }
-    public int Carbohydrates { get; set; }
-
+    public double Protein { get; set; }
+    public double Fat { get; set; }
+    public double Carbohydrates { get; set; }
     public string Category { get; set; }
 }
